Support for callsite discovery now has a default based on the backtrace() system call (in execinfo.h) if unwinders are not supplied at TAU configuration time. In the event backtrace is also not available, an attempt to use TAU_CALLSITE=1 will result in a warning message the first time it is used. The output from such a run will not generate callsite specific information.
What the default will not currently do is kick in if an unwinder fails.
Such support could potentially be added in the future.

Also added in this version, StackwalkerAPI's unwinder can be used
for callsite discovery. This has only been tested against GNU compilers
on x86_64 and is meant as a prototype for further development/testing.
It is NOT intended for supported production use.


Former-commit-id: 0435c4c1c97b0206625fdaeeb8de53930ed7e5a6