TauBfd.cpp should have the BGQ tag alongside the BGP tag when attempting to discover the location of the job's executable for the purposes of address-to-name resolution. Apparently the correct way of doing this for the BGQ is through the call Kernel_GetJobId() but it requires some special treatment (for inlining). Since what we have is not currently broken, we won't fix that until we have to.
Also updated is for a convenient way to build BGQ TAU using the GNU software stack. The key
for now is to point mpicxx to the gcc stack. For some reason, they now support mpicxx as a
valid wrapper for the default xlc environment (mpicxx used to be exclusive to the BGP).


Former-commit-id: f85af2c7ab5de5555c0c37d99d3df10307708cea