Updated the saveIntervalEvent() method to support TAUdb
The saveIntervalEvent() method will check the database schema
version and appropriately upload to the database.

Thing is, we need to refactor the IntervalEvent object entirely.
It should probably be a proxy object around a Function object,
or maybe an extension of one. Right now, when we save to the database,
we copy the Function objects to IntervalEvent objects, and then
save those to the database. It would be better if the IntervalEvent
object (at a minimum) had a reference to the underlying Function
object, so that no copying happens.


Former-commit-id: 927dd462cc90c2c0e0b368521728126f5f44ee58