Trying to fully metadata query support.
There is still a bug - when the Trial object is created (or
TAUdbTrial object, as the case may be), the DataSource for the
trial is not always set. Without it, things break. TAUdbTrial
objects should reference a TAUdbTrial object, and Trial objects
should reference a DataSource object. This seems to happen in
ParaProf, but not in PerfExplorer.


Former-commit-id: d98396223b2a7aaebe4c9163944dbf2d62e71eb4