Wow. OK, updated the TAUdb C API for new schema.
Added the primary metadata and supported the derived threads.
The PerfDMF support is now lagging in both metadata and
derived threads (mean and total are the only ones that exist
in the database). The TAUdb support lacks the secondary
metadata support, but that should be straightforward(?) to add.

Also added lots of test cases for testing the TAUdb schema.


Former-commit-id: ad511528b0308f65f8c2e1da608492e5a3b675ff