Implemented workaround for nasty internal BFD corruption when fed a bad address from a binary generated by the Intel 12.x compilers. This has been detected on at least three machines - French x86_64 machine at CEA, Mist (12.1 only) and ACISS (12.0 and 12.1 when it last worked).
The workaround currently introduces a non-trivial overhead paid at MPI_Finalize or any other
TAU finalization step. This is because we force BFD to re-initialize each time we attempt to
resolve an address to its symbol name. As a workaround, this is not necessary. Unfortunately,
we do not know the root cause for this internal corruption. If we do, then it is only
necessary to re-initialize BFD only when we encounter such a scenario, saving us from having
to pay some of that overhead.

If we do not find the root cause, then hopefully, the binutils team and/or Intel compiler
team will automagically find a solution and I can rid myself of this awkward and painful
workaround that had effectively cost me a nice sunny weekend.


Former-commit-id: 1b56ec4f1267d26db4d8e9865856106e6ccc78cf