For a variety of reasons, the code for the output to metadata of a backtrace was located in Comp_gnu.cpp, which for good reasons had its code base guarded by a ifndef for TAU_XLC. Unfortunately, this meant that XLC loses access to this functionality.
This refactor of the code allows XLC access, with its own requirements for
backtrace acquisition and name resolution. The code has been tested against
matmult on the BG/Q using both the XLC and GNU versions of TAU. These tests
are not perfect - for some reason, profile.0.0.0 refuses to be written when
call abort is used to trigger the signal. I will need to quickly test this
refactor against other platforms. For now, we will at least get something
for XLC for the BG/Q system.


Former-commit-id: cb0895bb6a30556710d41b81dfcf3d9c1a0a2cc6