Major refactor - AtomicEvent now only has static methods.
The AtomicEvent class was only used as a layover point for UserEvent
data going to or from the database. The database now directly creates
UserEvent objects, using the static methods in AtomicEvent.

The same refactoring needs to be done for AtomicLocationProfile,
IntervalEvent, and IntervalEventProfile objects. This will speed
up processing, because memory won't be wasted and data won't be
copied twice.


Former-commit-id: 53ee6ed5f0a0987d0bb8bf1758cf4e81a59f64ab