Preparatory work to allow PAPI 5.0 to work with TAU. The current pre-release version is 4.9. The interface to PAPI Components has changed. Some investigation is required to figure out what exactly needs to be changed.
This checkin only allows TAU to be built with PAPI 4.9 (the pre-release) but
with PAPI_COMPONENT_INDEX(a) set to return 0 (disabled). Once we figure out
the right way to use Components in PAPI, this will be replaced with the
functional code and there will be joy in the universe again.


Former-commit-id: 4a38b73c4020158b8d8d5544d51ca71175d4cbd4