Performance improvements (1000x!) for querying timer_callpath table.
The performance was so bad because this is a recursive query that
was NOT filtering by trial id. Therefore, the full callgraph of every
trial that was loaded into the database was constructed, then the
relevant values filtered out. That is now fixed to only work on
the relevant trial.


Former-commit-id: a38cfb218ad817453ddc235bc4c8df7d18d536cd