Wrapper modules NEED to have a proper "clean:" tag!
If missing, it will prevent our configure scripts from properly rebuilding
the wrapper module (in this case, pthread) when requesting a new TAU
configuration. This really should be documented for future developers of
TAU. The effects are very subtle, but also very wholly wrong.

As am example, this is the scenario I came up against:
1) Build TAU with -pthread (enables pthread wrapping) for the first time - works great.
2) Tested TAU with -cupti for the first time - works great (false impression).
3) Discovered some weird -cupti build artifacts, deleted <arch>, got a fix from Scott,
   rebuilt -cupti. - Now, everything went to hell because the old -pthread object files
   from #1 which did NOT go away (because there was no clean tag) happily formed the
   basis for a non-cupti .so file that the cupti version of TAU will eventually attempt
   to dynamically link.
The symptoms would lead one to believe early that it is anything but the final root cause.
This has been a painful time-consuming bug hunt and I'd like to spare future developers
from a similar waste of time.


Former-commit-id: c6c5d8cc551946ea207a54b1afd33820e3233cce