Fix for CUPTI issue. This does NOT address the still-unknown root cause!
Calls to TauTraceOneSidedMsg need to be guarded by checks for whether
TAU_TRACE is active.

The effect of not guarding these was exposed by QMCPack in the form of
unexpected trace files written out in spite of every part of TAU
reporting that tracing is NOT active! This does NOT happen with simple
GPU codes. As it stands, the root cause for this unintended trace output
is unknown. What is fairly clear, however, is that the fix is appropriate.
We should not be attempting to write to the trace buffers if tracing is
not active!


Former-commit-id: 4c5a0f1c76970e65bf475aeb10fb168cf1acd119