Fixed UPC array declarators undefined symbols bug.
According to source comments, ppw and gasp-dump use
gaspu.upc as a "magic linker flag" that causes the
UPC linker to act differently.  When this particular
source file is passed at link time, unresolved gaspu
symbols are "magically" resolved.

Sounds crazy, but it doesn't work to compile gaspu.upc
and then link the object!  I've only made this work
when passing the source itself.

This seems so broken that I won't be surprised if it
gets "fixed" in the near future.


Former-commit-id: 6b97789370dacb1d1ea4c1538601bac4c71470f3