Got rid of what I think is too stringent a check for TAU EBS timer interrupt support.
Fixed problem dealing with multiple metrics. Our current approach requires us to
read off a delta value at a sample (to handle arbitrary interrupt delivery on threads).
Unfortunately, this means the first sample has a 0 value as its previous value. This
is a problem for the TIME metric since the current timestamp is huge. The result was
giant delta values. My fix involves updating the base values from which deltas are
computed whenever we start a measured event. This way, when a sample tries to compute
its delta, it would be a better approximation to get the base from the start of the
last known measured TAU event than from the last known sample.


Former-commit-id: df901a5f3ae29f3ef1c8d6a54cb5f67348b6386d