Removed malloc call in signal handling.
The signal handling code has rather specific requirements. Basically,
any code that is called during the signal handler has to be signal-safe.
It also has to be non-reentrant. Therefore, calling "new" is BAD - it
eventually calls malloc.

I replaced the new call with a special memory allocation by the TAU
memory manager, and then used the placement new function to perform
the constructor code. The change is basically to replace this:

accumulator = new TauPathAccumulator(1,interval);

with this:

accumulator = (TauPathAccumulator*)Tau_MemMgr_malloc(1, sizeof(TauPathAccumulator));
new(accumulator) TauPathAccumulator(1,interval);

In addition, I make the samplingDisabled flag check more explicit in
the signal handling code. Therefore, we are sure to not process a signal
when we are already processing a signal.


Former-commit-id: d035cec593a39500d84835ef0907b841fb308cfe