Prevent crash with -PROFILECOMMUNICATORS and TAU_COMM_MATRIX
The code which adds the communicator to the MPI event assumes
that the parameter for the parameter based profile is the communicator.
That is not the case if TAU_COMM_MATRIX is defined at runtime,
because there are times where the parameter is "message size".
We now check the parameter name and process accordingly.


Former-commit-id: 3a8ca6579ea42609b4c5193a60f3e7f3b0dbf333