Prevent crash with -PROFILECOMMUNICATORS and TAU_COMM_MATRIX
The code which adds the communicator to the MPI event assumes
that the parameter for the parameter based profile is the communicator.
That is not the case if TAU_COMM_MATRIX is defined at runtime,
because there are times where the parameter is "message size".
We now check the parameter name and process accordingly.


Former-commit-id: 1ca57fc307de8f614335ebc48c98a4660c0b5c3e