Added code to read the XML_METADATA_GZ column.
Reading the binary data from the database works fine, but the
binary data is a GZipped string that needs to be uncompressed.
Zlib is having problems with it. Not sure what to do next.

After the string is expanded, it needs to be parsed and put
in the primary_metadata structure.


Former-commit-id: e2afaf346e6ea8e0d9998529a7410d48b59c441f