1. There was a verison of TAUdbTrial in the perfdmf package and the perfdmf.taudb package.  I refactored and merged the two.  The one in taudb did not inherit from Trial, so this caused many problems. It also shadowed private variables in Trial that it should have inherited instead.  (For example, trialID)
 2. TAUdbDataSource also had two different versions, one in
perfdmf.taudb and one in perfdmf.  In this case, the taudb version was a
wrapper around the data_source table.  (This table maps different types
of profiles to an id.)  This is very different from the
perfdmf.TAUdbDataSource class, which essentially the TAUdb version of
DataSource. So, I rename taudb.TAUdbDataSource to
taudb.TAUdbDataSourceType and moved perfdmf.TAUdbDataSource to the taudb
package.

 3. TAUdbSession objects are required for some of the methods in
TAUdbTrial to work; however, at the time the TAUdbTrials are created,
TAUdbSessions have not been created.  In fact, they are only created in
the main methods in the taudb package.  So, it would be best to
completely remove this class; otherwise, major refactoring work in
ParaProf, PerfExplorer and the rest of PerfDMF is needed to create and
store TAUdbSession objects.  Completely removing the class will also
take major refactoring work, so for now, I have basically made the
TAUdbSession a wrapper around a DB object.  This appears to be the only
part of TAUdbSession that is required for ParaProf/Perfexplorer/PerfDMF
to work.

 4. Fixed Null Pointer Exception when expanding Views with Sub-Views.

 5. Moved TAUdbDatabaseAPI to taudb. This made the
TAUdbDatabaseAPI(DatabaseAPI api) constructor invalid because it
accessed protect members of the passed in DatabaseAPI object.  To fix
this, I add the the copy(DatabaseAPI copy) to method to DatabaseAPI
which copies all of its variables to the passed in object.

 Note:  PerfExplorer and ParaProf display Views differently; I plan to
take a look at this after loading Trials in ParaProf is fixed.


Former-commit-id: 53208d2535ada40624ce30b74b606177d800a682