A couple of minor changes to fix merged profile ouput.
First, all threads were creating a .TAU application FunctionInfo
object, so each thread had a unique top level timer, which is bad.
The FunctionInfo object has per-thread data storage internally.
This is especially bad for callpath profiles, because the timers
are duplicated in the header, although they have the exact same
name.

Second, threads which do not call a function should not output
zeros in the interval_data section. This both makes the profile
larger, but also confuses ParaProf, which interprets those zeros
as meaningful data. This is particularly bad in the callpath
tree table view, with lots of zero timers.


Former-commit-id: 85f8ed34fb6e040a525f5ca8bc2d85599a38daa5