Major changes to locking for OpenMP for top level timer creation.
The "create top level timer if necessary" code calls the pure
timer creating code, which requires the thread ID. Well, the top
level timer is created when the thread is registered, and the
thread ID doesn't exist yet. So, we added a new, special lock to
the OpenMP layer, just for the thread registration. Then, we call
threadLockEnv(), rather than LockEnv(), so that we avoid any calls
to myThread() before TAU knows the thread ID.


Former-commit-id: 357ea2012616de04bb0d6a10e7d50fb49b0469e3