When a user specifies -scalasca=<dir> using Scalasca-2.x (which doesn't have
its own measurement layer), TAU checks the path for an appropriate Score-P and uses -scorep=<dir> informing the user. If one is not found, it exits prompting the user to specify a new Score-P dir. You may still use scan with the executable generated by TAU and use paraprof etc.


Former-commit-id: 2d96df0eb56b95a6bc956ca18ad5a3bde358b141