Hacky thread-unsafe work around for OpenMP
The horror, the horror...

A thread doesn't know if it's inside TAU code without getting its thread ID.
To get the thread ID, a thread may (probably will) call malloc.
The memory wrapper needs to know if the call is inside TAU, so it needs the thread ID.
To get the thread ID, a thread may (probably will) call malloc, and so on.

TAU *needs* a way to discover if it is profiling itself without relying on the thread ID.


Former-commit-id: 828110a8c08a1dc0fdf3686763fc5d65d3b4f4eb