Multithreaded memory wrapper.
Instead of implicitly disabling instrumentation by setting insideTAU for
a given thread ID when entering TAU, explicitly disable instrumentation
for a specific thread when entering TAU.

This explicit approach works even when a thread is not fully initialized
so we can profile the internals of thread creation (i.e. clone()).  It
is faster and more efficient since RtsLayer is not queried on every
memory (de)allocation.  There is little or no overhead in single
threaded applications.

This needs more testing, but it's working well with GNU at least.


Former-commit-id: 50a035187fea810c9456f73e0d92a4cf970e67bb