Committing changes for the Collector API and Sampling.
The collector API now has a way to track when worker threads enter the
parallel region. This is done by wrapping the worker thread function
with a proxy function, which calls the collector API events before
and after the thread executes the function pointer.

We shouldn't see any more race conditions in the collector, which is awesome.

There was also a bug in how threads were identified.
We were using omp_get_thread_num() in both the GOMP wrapper and in
the collector API, and it should use the TAU thread ID instead,
all the time.

Finally, the sampling now resolves to both the function and to the line.
This allows for aggregation of samples within a function to show the total
time sampled in a function.


Former-commit-id: 43ec00f4ecd38e7ce9c3b5ea650fb1a0d6563f87