Most importantly, fixed bug where .TAU applicaiton was getting created multiple times when initializing sampling with OpenMP.
There is a parallel region which has a critical section to
create initialize sampling, which includes creating a top level
timer. Unfortunately, all locking has to be released by thread
0 before this region, or there is deadlock. Therefore, I changed
the parallel region to an ordered region, guaranteeing that
thread 0 (who used to have the lock) goes first, and no other
thread can acquire the DB lock until thread 0 is done with it.

Also, some new niceties in the collector API, although they are disabled for now.


Former-commit-id: ad10f332e63d1fdf886032585b8027ba96c575de