Fixing configure so that BFD and libunwind can configure with Apple. It doesn't work at runtime (all BFD samples are unresolved), but at least the configure works.
Also fixed conigure for Apple when a non-standard gcc is used.
for example, you can download and build gcc 4.7, and it works,
and unlike the gcc included with Apple it supports OpenMP. However,
it does not take the -rdynamic nor -Wl,-export-dynamic options.

One caveat with BFD and libunwind support on apple - it requires
that you install them with MacPorts (the configure assumes that
you are using the /opt/local/lib and inclue directories).


Former-commit-id: b83aeb149107cd04a651225eeca18a3037ab0334