After a new node is created in Comp_gnu (HashNode), it is not initialized. This led to a very elusive bug. When we used -optCompInst -optShared, it appeared at LLNL but not on Aciss. When -optShared was not used, -optCompInst seemed to work properly. We were doing:
node = new HashNode;

...
if (!node->fi)
  call TAU_PROFILER_CREATE and get a new fi.
But sometimes it would have uninitialized garbage (when -optShared was used).
So, the TAU_PROFILER_CREATE would not get called and TAU was crashing inside a
fi->GetProfileGroup() call because fi was not a valid address. We fixed this using

node = new HashNode;
node->fi = NULL;


Former-commit-id: de6780003ad1596ccb57d11ea6f57b8d89d41095