Integrated GOMP wrapper into TAU.
At configure, TAU will determine if GCC and -openmp are used. If so,
the gomp wrapper will be built. When using PDT, the symbols are
wrapped at link time.

Without instrumentation, there is a new option to tau_exec. The option
is -gomp, and it will preload the GOMP wrapper. When configured with
libunwind and bfd, all parallel region information is available.


Former-commit-id: 87748d98f98229521d1e2a72de4557f3fcb5106d