Initial support for OMPT, only tested with Intel.
If you want to try it, here are the instructions:

I haven't tested this with the IBM compiler, but this tested with the Intel compiler. I assume that the IBM experimental compiler will have the OMPT built in, so you won't need the instructions to build the ompt support outside of TAU. You also shouldn't need to modify the makefile to build the example - you should be able to add the "-useropt=-DTAU_USE_OMPT" to the TAU configure. But I am not sure.  The following instructions are for building it on manticore (or mist, or aciss, or tank, etc.) with the Intel compiler (13.1.0 is the only version I tested).

to build and test with Intel (this worked on manticore):
1) module load intel/13.1.0
2) Get the Intel OpenMP library from googlecode svn:
svn checkout http://intel-openmp-rtl.googlecode.com/svn/branches/ompt-support ompt-support
cd ompt-support/itt/libomp_oss
make (the default configuration should work with intel compilers just fine.)
note: the library will be in the directory tmp/lin_32e-rtl_5_nor_dyn.12.rel.40.c0.t0-manticore/libiomp5.so
3) Get the latest TAU from git.
4) configure with:
./configure -pdt=<pdt-path> -pdt_c++=g++ -openmp -bfd=download -cc=icc -c++=icpc -fortran=intel -useropt=-DTAU_USE_OMPT
5) make install
6) cd examples/mm
7) Add this to the link line:
 -Wl,-rpath=<path-to-libiomp5.so> <path-to-libiomp5.so>/libiomp5.so
 8) make and run, you should see parallel regions and barriers.

Former-commit-id: 1b817ba0cf0fdc359a961c9af5d0d351a2917ae3