GOMP,Sampling: enhancements, bug fixes, optimizations.
I have fixed sampling so that on systems that support it (i.e. Linux),
a per-thread wallclock timer is used for sampling, which enables samples
during idle periods.

Also, the sampling measurements were incorrect, in that they were being
modified if a timer started during the sample period. I have disabled that
behavior for now, it should be removed permanently with further testing.

The GOMP wrapper is fixed, in that it now measures locks, and it works
wether you link with TAU or use tau_exec.


Former-commit-id: aa8faffac42de2cdcd07e2fa54db5a1869febbaa