Thread local storage for PGI!
Thread local storage (TLS) is not easily supported with some compilers,
like PGI. TLS global variables are specified this way:

__thread int foo;  // every thread has their own foo copy
or
__declspec(thread) int foo; // every thread has their own foo copy

Without that, we have to rely on pthread_(g/s)etSpecific() calls.
The OpenMPLayer.cpp and TauCAPI.cpp files now have a shared object
for thread specific global data, and that object is in the TauPthreadGlobal.cpp
class. This is the same way the pthread layer supports thread ids,
so this is nothing new in TAU, just that we are applying it to the OpenMP
thread management. And we are applying it for the "insideTAU" support.
In benohmarks, this reduced execution time of NPB 3.2.1 BT.W from 13 to 10.25 seconds
on ACISS with the PGI 13.6 compiler on one node (throttling disabled to increase overhead).


Former-commit-id: e84f6f2ebd06a126484504fafc5dec9d795fc5ca