Cleanup running timers after throttling.
When a timer is throttled on one thread, it might still be
running on other threads. The stop event will be ignored, because
the event is now throttled. When that happens, the next "Tau_stop_current_timer()"
call will fail, because the *throttled* event will be the one that
gets stopped. this breaks the collector API support.

Instead, when stopping the "current" timer, check to see if the
current timer(s) are throttled, and if so, stop them and then stop the
next timer on the stack (the one the user actually wanted stopped).


Former-commit-id: ea29d2d731ecc1345eb86c14d66d7d1366e7e73d