Fixed callsite profiling.
There were two problems. First, when the global stack of Profiler
objects is created, they are allocated into uninitialized memory.
There are several checks which check for NULLs in the object, they
can fall through with garbage data. Second, the timer name was not
getting correctly assigned - the snprintf function was only printing
8 bytes (the sizeof(ptr), which is 8 bytes!).


Former-commit-id: 70a6dac789efa22f81af3504e2cbc46b91f14470