Reallocating the event stack results in corrupted memory.
The code has been modified to support multiple scenarios, either with
realloc or malloc with memcopy. Neither seems to work correctly all the time.
They are both implemented here. In case the memory does get corrupted, it
can be "restored" when timers are stopped - see the code in Profiler.cpp
with the following comment:

+/* It is possible that when the event stack gets deep, and has to be
+ * reallocated, the pointers in the event stack get messed up. This
+ * fixes the parent pointer for flat profiles, but I don't know if it
+ * is a robust fix for all scenarios! - Kevin */


Former-commit-id: f95629bf2a57dd84b463818e296cfaa4fc1f77fe