AAAAAAAAAAAAAARRRRRRRRRRRRRRGGGGGGGGGGGGGHHHHHHHHHHHHH!!!!!!!
The TauMemory.cpp code wasn't incrementing the "inside TAU"
counter when checking to see if memory was allocated by TAU or the
system when it calls the "Find" method. Now that the GOMP wrapper
is incr/decr that counter, memory wrapping was getting locked.
Why? Well, one of the SIDE EFFECTS of incr/decr is
enabling/disabling the memory wrapper.  Anywho, the deadlock
is now fixed.


Former-commit-id: 99f8ec3803bc0a4d96deb424aedd6e8658c3b1ec