asmdex: new patch to fix a lable rewriting bug
Some instruction needs to be modified after you change other portion
of the bytecode. For example, goto may need to be changed to goto/16
because the offset cannot be put into 1 byte anymore after you insert
some additional instructions. So asmdex will check and correct all
the label references when necessary in MethodWriter::visitEnd().

Unfortunately, there is a subtle bug which will keep the swith table
offset of packed-switch/sparse-switch unfixed. This patch fixes this
bug.


Former-commit-id: 405a9238714a8b09ecd49845c857e7a0317a07c9