Minor bug fix: minus after bitwise-or
In order to calculate the padding length, my intention is to first
bitwise-or then minus. I forgot the parentheses and actually minus
has the higher precedence than bitwish-or, so this is actually a
bug. I notice this because compiler gives a warning.

It is interesting that actually they compute the same thing with or
without the parentheses...


Former-commit-id: 4297d15d13019547bb976dfa8b2f80af29281cb5