Added -Wl,-rpath,<mpilibdir> to the TAU_MPI_LIB and TAU_MPI_FLIB variables in include/Makefile. This is needed for power profiling.
Also, in Score-P we were seeing twice the number of marker events. This is fixed by turning off context events in Score-P because the data structure used in Score-P is that of a tree. So, a user-event is just as effective as a context event. We do not need to trigger it twice as we do in usual TAU profiles.


Former-commit-id: f4053c482c6917b48b646e3b2b1692d271cea40b