dexInjector: fix register mapping for method parameters in debug info
Name and desc of method parameters are not included in debug info
as they can be extracted from method descriptor. As we have extended
register space of the method, we shall fix the register allocation
for method parameters recorded in debug info.

If you run "dexdump -d" on injected dex file without this fix, most
likely you will see dexdump complains like this:

   E/dalvik: Invalid debug info stream.....


Former-commit-id: d4f578c9bfc2089b77636c52ac1c88811edb189d