dexInjector: fix string index overflow issue
We pass the callsite as a string constant to Profiler.start()/stop().
As a result, we are increasing strings in dex constant pool. Chances
are that index of some strings in constant pool will be bigger than
0xffff, i.e. can not be hold in 2 bytes.

INSN_CONST_STRING is using 2 bytes to hold the string index, thus
it should be rewritten to INSN_CONST_STRING_JUMBO if the string index
it is referring does not fit in 2 bytes.

ASMDEX hides the string index from us so this fix should be done by
ASMDEX. The fix is not trivial so instead we workaround this in our
code by blindly rewrite all INSN_CONST_STRING to its JUMBO version.


Former-commit-id: 23c068cba63776e9f9f93aa2c321ad26a3d46c22