Fix building with Python on OSX.
For reasons unknown to me, configuring with -python builds TAU as a
Darwin bundle (aka dynamic module) instead of a dynamic library.
Bundles cannot be linked against!  They must be opened with dlopen().
As far as I can tell, you will never want libTAU.so to be a bundle.

If anyone can tell me why libTAU.so should be a bundle I'll reverse
this commit :-)


Former-commit-id: 34e19fef02696ba712a6a178d378cc475669a833