Reinstate strdup leak
We can't assign to a char * that isn't reallocated.
Name and group name are allocated by accident when a
FunctionInfo is created, but assigning to this pointer
later causes further leaks and will segfault if the
function name length grows, i.e. when [THROTTLED] is added.

Yes, this is a leak, but overwriting the pointer doesn't fix it.

Long term, we really need to fix string abuse throughout TAU.


Former-commit-id: 2dd549985a19df8791512784951d1a4ad55e7cc7