Handle Exceptions in JVMTI.
We were not handling exceptions. This meant that if an
exception was thrown causing us to return to the caller,
the method's timer would not be stopped and we would get
an overlapping timer error when the caller returned.

We now stop the timer of a method if it throws an
exception which is not caught in the same method
(that is, if it causes control to pass to another method).


Former-commit-id: 8de1f1eb641d6d77e7dfef5c24eb4f6caf173ae8