Implement TAU_CALLSITE_DEPTH.
There was a TAU_CALLSITE_LIMIT that did absolutely nothing since
that value was never checked.  Now it's called TAU_CALLSITE_DEPTH
to be consistent with TAU_CALLPATH_DEPTH and has a similar effect.
TAU_CALLSITE_DEPTH=10 will unwind the callstack up to 10 callsites
and generate callsite events for each.


Former-commit-id: b962585c744e0b90bc9fe31d264abc55dcde5bbe