Implement TAU_CALLSITE_DEPTH.
There was a TAU_CALLSITE_LIMIT that did absolutely nothing since
that value was never checked.  Now it's called TAU_CALLSITE_DEPTH
to be consistent with TAU_CALLPATH_DEPTH and has a similar effect.
TAU_CALLSITE_DEPTH=10 will unwind the callstack up to 10 callsites
and generate callsite events for each.


Former-commit-id: fd8008d42b544d90ce4e7d96e663d3045ac8d277