Initial implementation of PySpark profiling
Uses Spark's built-in python profiling interface to profile
PySpark tasks with PyTAU. Currently, sets the node ID for
each task to the task attempt ID, which is not ideal (tasks are
more like threads than nodes) but does work.

This requires Spark 2.2 or later, as the task attempt ID is not
available from Python in earlier versions. Note that Spark 2.2
has not actually been released yet, so, for now, this requires
the git HEAD of Spark.


Former-commit-id: 1903392088b1112772abcea0778308b246705092