tau_spark-submit wrapper for Spark
To ensure that TAU environment variables are properly set in the
driver and in all workers, the tau_spark-submit script can be used
instead of spark-submit. The wrapper will create a temporary copy
of your existing Spark configuration files, if they exist, and add
TAU-specific environment variables to its copy of spark-env.sh.


Former-commit-id: 3885bc69203b9ba38e3e1a08dab24c4f9fc858e8