Fix symbol resolution with non-zero vaddr
Symbols were not resolving for sampling or compiler-based
instrumentation for certain shared libraries on the ERDC Onyx system.
This was caused by the text section of the executable being loaded at a
non-zero offset into virtual memory, which TAU's BFD support was not
handling. TAU now examines the ELF header (if support in BFD is
available) to determine the offset and tries (addr + offset) if
resolution otherwise fails.

Squashed commit of the following:

commit 17bf1de1cf4bd80ccba0078cbf2ae1c63542e857
Author: Nicholas Chaimov <nchaimov@paratools.com>
Date:   Fri Apr 6 16:23:23 2018 -0700

    Use offset in executable as well

commit 56b12d84289734fe6b4bdbcc1a04f5d87033c262
Author: Nicholas Chaimov <nchaimov@paratools.com>
Date:   Fri Apr 6 16:14:06 2018 -0700

    Work towards fixing symbol resolution on Onyx


Former-commit-id: 6b2991241230ee326950a9da118bd72dcb4d6f0c