By default, when an MPI configuration is used with a serial application, we get a message such as: TAU: WARNING! An MPI configuration was used in TAU, but MPI_Init was not called. No data will be written for pid=72541.
and no profile data is written. We have now introduced an option:

export TAU_SET_NODE=<node_number>
that will forcibly set the node number for MPI (only) applications. So, while
configuring HDF5, if it runs a set of tests, the tests will not add this message to the output of say H5detect which creates files that are compiled subsequently.
There are other cases (e.g., serial Python run with TAU configured with MPI in ECP SDK image), where being able to force the creation of TAU profiles and traces using TAU_SET_NODE=0 may be useful.


Former-commit-id: 066ffa6936dfdd63ba85229b9e53f0f5b3e410bd