Add function->line resolution with libdwarf
libbfd is ***VERY*** slow at repeatedly resolving the line number of a
function. It appears that it reparses the entire debug symbol table with
every lookup. This commit will now use libdwarf instead, if available, to
get the line numbers for -ebs_resolution=function. To enable, build with
both -bfd=download and -dwarf=download.


Former-commit-id: bbf7e117f0f5ffac613c01b251311e696c2b077a